ServerPython/requirements.txt
requests==2.28.1
wordcloud==1.8.2.2  
numpy==1.23.5
matplotlib==3.6.2
Pillow==9.3.0
Flask==2.2.2